Documents/GitHub/sigmatic/neutral-strat/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Data and API
python-binance>=1.0.16
requests>=2.28.0
websockets>=10.0

# Configuration
PyYAML>=6.0
python-dotenv>=0.19.0

# Backtesting and analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Database (optional)
redis>=4.3.0
sqlalchemy>=1.4.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.8.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.971

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

# Utilities
click>=8.0.0
tqdm>=4.64.0
schedule>=1.1.0

# Monitoring
prometheus-client>=0.14.0